requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
numpy
pandas
PyYAML
psycopg2-binary